backend/embedding-service/requirements.txt
fastapi
uvicorn
langchain
langchain-chroma
langchain-openai
chromadb
python-dotenv